Started some work on abilities
Added a new file ability_dictionary.py, which builds a dictionary
mapping strings to functions. The strings are the name of the ability
and the function is executable code that changes the game state in some
way. Currently, there is only one ability, 'barrier', which creates an
immovable impassable barrier token.

Also made it optional for game.depoy to check legality, which makes
things like creating tokens and setting up games much easier.